refactor: reorganize component folder structure
- Move chat components to components/chat/
- Move UI components to components/ui/
- Convert .js files to .tsx for TypeScript
- Update all import paths
- Use index.tsx pattern for component exports
- Improve folder organization by feature
