+++ __tests__/ChatWindow.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { ChatWindow } from '../components/ChatWindow';
+import { ChatWindow } from '../components/chat/ChatWindow';
 import type { Message } from '../types/Message';
 
 describe('ChatWindow', () => {
