+++ app/chat/page.tsx
@@ -3,12 +3,12 @@
 import styles from './ChatPage.module.css';
 import { useState, useEffect, useRef } from 'react';
 import Image from 'next/image';
-import UserIcon from '../../components/Icons/UserIcon';
-import ArchiveIcon from '../../components/Icons/ArchiveIcon';
+import UserIcon from '../../components/ui/Icon/UserIcon';
+import ArchiveIcon from '../../components/ui/Icon/ArchiveIcon';
 import { processUserMessage, summarizeChat } from '../LLM/LLMService';
-import { Navbar } from '../../components/layout/Navbar';
-import { NavIconButton } from '../../components/Buttons/NavIconButton';
-import { ChatWindow } from '../../components/ChatWindow';
+import { Navbar } from '../../components/ui/Layout/Navbar';
+import { NavIconButton } from '../../components/ui/Button/NavIconButton';
+import { ChatWindow } from '../../components/chat/ChatWindow';
 import { Message } from '../../types/Message';
 
 export default function ChatPage() {
