+++ components/ui/Icon/RobotIcon/index.tsx
@@ -1,3 +1,4 @@
+import React from 'react';
 import styles from './RobotIcon.module.css';
 
 export default function RobotIcon() {
