+++ components/ui/Icon/UserIcon/index.tsx
@@ -1,3 +1,5 @@
+import React from 'react';
+
 export default function UserIcon() {
   return (
     <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" width="24" height="24">
